==> notebooks/01a_Data_Cleaning_radiomics.ipynb <==
# Reformat and reorganize the radiomics data

Assumed data organization:
- $PATH/
    - raw/ (contains input data files)
    - tables/ (contains output data files)
    - figures/ (contains output image files)

# External packages
import pandas as pd
import pickle
import yaml
import os
import sys

# Local packages
sys.path.append('../src')
from ninetynine import ninetynine

# Read in data
with open('../config/paths.yaml','r') as file:
    paths_list = yaml.safe_load(file)
    PATH = os.path.abspath(paths_list['PATH'])

data = pd.read_csv(os.path.join(PATH,'raw','DL-Lung-Features-SARS-CoV-2-NHP-064-1-2-3_complete.csv'))

data.head()

# Get rid of space in column names, annoying to deal with later
data = data.rename(columns={'Study Day':'StudyDay'})
data = data.rename(columns={'Study Date':'StudyDate'})

# Get lists of the id columns and the dependent variable columns
id_cols=['Study','Subject','StudyDay','StudyDate','File Name']

##need to adjust lists with the .1 and .2 notation
shape_list=['VoxelVolume','Maximum3DDiameter','Compactness2','MeshVolume'
,'MajorAxisLength','Sphericity','LeastAxisLength','Elongation'
,'Compactness1','SurfaceVolumeRatio','Maximum2DDiameterSlice'
,'Flatness','SurfaceArea','MinorAxisLength'
,'Maximum2DDiameterColumn','SphericalDisproportion','Maximum2DDiameterRow']

glcm_list=['JointAverage','Autocorrelation','JointEntropy','ClusterShade',
'MaximumProbability','Idmn','JointEnergy','Contrast',
'DifferenceEntropy','InverseVariance','DifferenceVariance',
'Idn','Idm','Correlation','SumAverage',
'SumEntropy','MCC','SumSquares','ClusterProminence',
'Imc2','Imc1','DifferenceAverage','Id','ClusterTendency']

gldm_list=['GrayLevelVariance','HighGrayLevelEmphasis'
,'DependenceEntropy','DependenceNonUniformity'
,'GrayLevelNonUniformity','SmallDependenceEmphasis'
,'SmallDependenceHighGrayLevelEmphasis','DependenceNonUniformityNormalized'
,'LargeDependenceEmphasis','LargeDependenceLowGrayLevelEmphasis'
,'DependenceVariance','LargeDependenceHighGrayLevelEmphasis'
,'SmallDependenceLowGrayLevelEmphasis','LowGrayLevelEmphasis']

firstorder_list=['InterquartileRange','Skewness'
,'Uniformity','Median','Energy'
,'RobustMeanAbsoluteDeviation','MeanAbsoluteDeviation'
,'StandardDeviation','TotalEnergy'
,'RootMeanSquared','90Percentile'
,'Minimum','Entropy','Range'
,'Variance','10Percentile','Kurtosis'
,'Maximum','Mean']

glrlm_list=['ShortRunLowGrayLevelEmphasis','GrayLevelVariance.1'
,'LowGrayLevelRunEmphasis','GrayLevelNonUniformityNormalized'
,'RunVariance','GrayLevelNonUniformity.1'
,'LongRunEmphasis','ShortRunHighGrayLevelEmphasis'
,'RunLengthNonUniformity','ShortRunEmphasis'
,'LongRunHighGrayLevelEmphasis','RunPercentage'
,'LongRunLowGrayLevelEmphasis','RunEntropy'
,'HighGrayLevelRunEmphasis','RunLengthNonUniformityNormalized']

glszm_list=['GrayLevelVariance.2','ZoneVariance'
,'GrayLevelNonUniformityNormalized.1','SizeZoneNonUniformityNormalized'
,'SizeZoneNonUniformity','GrayLevelNonUniformity.2'
,'LargeAreaEmphasis','SmallAreaHighGrayLevelEmphasis'
,'ZonePercentage','LargeAreaLowGrayLevelEmphasis'
,'LargeAreaHighGrayLevelEmphasis','HighGrayLevelZoneEmphasis'
,'SmallAreaEmphasis','LowGrayLevelZoneEmphasis'
,'ZoneEntropy','SmallAreaLowGrayLevelEmphasis']

ngtdm_list=['Coarseness','Complexity','Strength'
,'Contrast.1','Busyness']

# Confirm there aren't any repeats
rep_bool = len(set(shape_list+glcm_list+gldm_list+firstorder_list+glrlm_list+glszm_list+ngtdm_list)) != len(shape_list+glcm_list+gldm_list+firstorder_list+glrlm_list+glszm_list+ngtdm_list)
ninetynine(rep_bool,'repeat features in grouping lists')

# Rename radiomic features to include class
for name in data.columns:
    if name in shape_list:
        data.rename(columns={name:name.split('.')[0]+'-shape'},inplace=True)
    elif name in glcm_list:
        data.rename(columns={name:name.split('.')[0]+'-glcm'},inplace=True)
    elif name in gldm_list:
        data.rename(columns={name:name.split('.')[0]+'-gldm'},inplace=True)
    elif name in firstorder_list:
        data.rename(columns={name:name.split('.')[0]+'-firstorder'},inplace=True)
    elif name in glrlm_list:
        data.rename(columns={name:name.split('.')[0]+'-glrlm'},inplace=True)
    elif name in glszm_list:
        data.rename(columns={name:name.split('.')[0]+'-glszm'},inplace=True)
    elif name in ngtdm_list:
        data.rename(columns={name:name.split('.')[0]+'-ngtdm'},inplace=True)
    else:
        print('Column not in list:',name)

# Feature name list for storage
var_cols=data.drop(id_cols+['Class'],axis=1).columns.tolist()

# Print column names
i=0
for name in data.columns:
    i += 1
    print (str(i)+') '+str(name))

### Drop entries with errors

# # Remove a specific set of entries that have erroneous StudyDay values
data.loc[(data.Subject == 'B03942') & (data.StudyDate < '2020/06/01')] = None
data.dropna(how='all',inplace=True)
# Fix specific errors in group coding
data.loc[data.Subject == 'B03942','Class'] = 'Mock'
data.loc[data.Subject == 'G57L','Class'] = 'Mock' 
data.loc[data.Subject == 'G57N','Class'] = 'Infected' 

# Entries where subject ID does not match filename (confirmed we should keep these)
mask = (data.Subject != data['File Name'].str.split('_', n = 1, expand=True)[0])
data.loc[mask,['Subject','File Name','StudyDay','Class']]

Remove entries at baseline or after day 8 time points

# Get a list of subjects that had a StudyDays 2, 4, and 6
subj_list = data.loc[data['StudyDay'].isin(['2','4','6']),'Subject']

# Drop all other subjects and extra rows of other time points (for simplicity)
ignore_days = ['BL','BL4','10','12','19','30']
data.drop(data[(data['StudyDay'].isin(ignore_days)) | (~data.Subject.isin(subj_list))].index, inplace = True)

# Check for rows with the same subj_id and StudyDay
ninetynine(len(data.loc[data.duplicated(subset=['Subject','StudyDay'])])>0,'rows with the same subj_id and StudyDay')

# Separate time points into separate columns
data = data.set_index(['Subject','StudyDay']).unstack()

Calculate the change from pre-exposure

for var in var_cols:
    # Collapse pre-exposure values (should have one per subject)
    data.loc[:,(var,'pre')] = data.loc[:,(var,['-5','-6','-7'])].mean(axis = 1, skipna = True)

    # Calculate the change from pre-exposure
    data.loc[:,(var,'pre_delta')] = data.loc[:,(var,'pre')]-data.loc[:,(var,'pre')] #should be all 0
    data.loc[:,(var,'2_delta')] = data.loc[:,(var,'2')]-data.loc[:,(var,'pre')]
    data.loc[:,(var,'4_delta')] = data.loc[:,(var,'4')]-data.loc[:,(var,'pre')]
    data.loc[:,(var,'6_delta')] = data.loc[:,(var,'6')]-data.loc[:,(var,'pre')]
    data.loc[:,(var,'8_delta')] = data.loc[:,(var,'8')]-data.loc[:,(var,'pre')]


data.head()

# Copy the class to the new columns from the original columns

# stack and unstack so that missing columns fill in with null
data = data.stack().unstack()

time_points=['pre','2','4','6','8','pre_delta','2_delta','4_delta','6_delta','8_delta']
delta_time=['pre_delta','2_delta','4_delta','6_delta','8_delta']
# back-fill and foward-fill incase the column order changes
data.loc[:,('Class',time_points)] = data.loc[:,('Class',time_points)].fillna(method='ffill',axis=1).fillna(method='bfill',axis=1)

# Reformat for readability
data = data.stack()

### View the data

# Uncomment for different popular views of the data
#data#.describe()
#data.unstack()['Strength']

#data.groupby(['StudyDay','Class']).describe()

#data.loc[data['StudyDay'] == -6]
#data.reset_index().loc[data['Subject'] == 'B03942',['Subject','StudyDay','Class']]
#data.loc[(data.Class == 'Mock') & (data.StudyDay == '2')]
#data.loc[data['Class'] == 'Infected',['Subject','Class']]
#data.reset_index()
#data.describe()
#data.loc[(data.Class == 'Mock') & (data.StudyDay == '2'),'Subject']

### Save resulting tables

# Save full table (no error values)
data.to_csv(os.path.join(PATH,'tables','data_radio.csv'))

# Save simplified table (only delta timepoints)
data = data.reset_index()
data_deltaOnly = data.drop(data[data['StudyDay'].isin(['-5','-6','-7','2','4','6','8','pre','pre_delta'])].index)
data_deltaOnly.set_index(['Subject','StudyDay']).to_csv(os.path.join(PATH,'tables','data_radio_delta.csv'))

### Exclude variables based on domain-specific feature screening criteria

var_MCExclude=['SurfaceVolumeRatio-shape','Maximum2DDiameterSlice-shape'
,'MaximumProbability-glcm','DifferenceEntropy-glcm'
,'InverseVariance-glcm','SmallDependenceEmphasis-gldm'
,'DependenceNonUniformityNormalized-gldm','LargeDependenceEmphasis-gldm'
,'LargeDependenceLowGrayLevelEmphasis-gldm','DependenceVariance-gldm'
,'SmallDependenceLowGrayLevelEmphasis-gldm'
,'Minimum-firstorder'
,'RunVariance-glrlm','LongRunEmphasis-glrlm'
,'ShortRunEmphasis-glrlm','RunPercentage-glrlm'
,'LongRunLowGrayLevelEmphasis-glrlm','RunLengthNonUniformityNormalized-glrlm'
,'ZonePercentage-glszm','LowGrayLevelZoneEmphasis-glszm'
,'SmallAreaLowGrayLevelEmphasis-glszm']

# Check that listed variables are all in the dataset
for var in var_MCExclude:
    if not var in data.columns.tolist():
        print('Not in dataset:',var)

# Get a list of subset of variables after domain-specific feature screening
var_cols = [ var for var in var_cols if not var in var_MCExclude]

# Write to status message file, change this anytime this part of the analysis changes
status='radioExcl_MC;'
with open(os.path.join('..','config','analysis_status','u_radio.txt'),'w') as out_file:
    out_file.write(status)

#### Save variables

# Saving the objects:
with open('../config/lists_radio.pkl', 'wb') as f:
    pickle.dump([id_cols,var_cols,time_points,delta_time], f)

# Print column names
i=0
for name in var_cols:
    i += 1
    print (str(i)+') '+str(name))
