==> notebooks/01b_Data_Cleaning_clinPath.ipynb <==
import pandas as pd
import pickle
import yaml
import os
import sys

# Local packages
sys.path.append('../src')
from ninetynine import ninetynine

#open data
with open('../config/paths.yaml','r') as file:
    paths_list = yaml.safe_load(file)
    PATH = os.path.abspath(paths_list['PATH'])

data_path = pd.read_csv(os.path.join(PATH,'raw','LW_Data_64e_wBL.csv'))
data_path.head()

# Views of the raw data
#data_path.astype('object').describe()
#len(data_path[data_path.NAME == 'C-Reactive Protein'])
#data_path.loc[data_path.NAME == 'C-Reactive Protein',['SAMPLE_NUMBER','SUBJECT_ID','STUDY_DAY','STUDY_NO']]
#data_path.loc[data_path.NAME == 'ICT',['SAMPLE_NUMBER','SUBJECT_ID','STUDY_DAY','STUDY_NO','ENTRY']].astype(object).describe()
#data_path.info()
#data_path.loc[data_path['SUBJECT_ID'] == 'B03757','NAME']
#data_path.loc[(data_path['SUBJECT_ID'] == 'G21E') & (data_path['NAME'] == 'Hemoglobin')]

Some subjects have errors on these entries (StudyDay does not match SAMPLED_DATE)\
However some subjects need this data (ex. B0757)\
Solution is to change STUDY_NO to something other than HOLDING in the specific entries we need to keep

### Rename columns and values

# Rename columns and values to follow previous convention
data_path.rename(columns = {'SUBJECT_ID':'Subject','CLASS_DESCRIPTOR':'Class','STUDY_DAY':'StudyDay'},inplace=True)
data_path.loc[data_path.Class == 'Virus','Class'] = 'Infected'

# Get rid of spaces and slashes, these are annoying later
data_path['NAME'] = data_path['NAME'].replace(' ','_',regex=True)
data_path['NAME'] = data_path['NAME'].replace('/','v',regex=True)

# Calculate age
data_path['Age_d'] = (pd.to_datetime(data_path.SAMPLED_DATE) - pd.to_datetime(data_path.DOB)).dt.days

### Remove error values

# Confirmed with SB that this is reasonable
data_path.loc[(data_path.NAME == 'C-Reactive_Protein') & (data_path.ENTRY == '<0.1'),'ENTRY'] = 0
data_path.loc[(data_path.NAME == 'Total_Bilirubin') & (data_path.ENTRY == '<0.1'),'ENTRY'] = 0
data_path.loc[(data_path.NAME == 'Alanine_Aminotransferase') & (data_path.ENTRY == '<10'),'ENTRY'] = 0
data_path.loc[(data_path.NAME == 'Triglyceride') & (data_path.ENTRY == '<10'),'ENTRY'] = 0
data_path.loc[(data_path.NAME == 'Creatinine_Kinase') & (data_path.ENTRY == 'Not enough sample'),'ENTRY'] = None

# Remove all values from the HOLDING protocol
    # Some subjects have errors on these entries (StudyDay does not match SAMPLED_DATE)
    # However some subjects need this data (ex. B0757)
    # Solution is to change STUDY_NO to something other than HOLDING in the specific entries we need to keep
# For specific subjects, save them by assigning a ney study procotol name
data_path.loc[(data_path['Subject'] == 'B03757') & (data_path['STUDY_NO'] == 'HOLDING'),'STUDY_NO'] = 'SAVE_WC'
# Remove remaining values on HOLDING protocol
data_path = data_path.loc[data_path.STUDY_NO != 'HOLDING']

# Remove additional timepoints, confirmed that these three subjects have other baseline values to pull from
data_path = data_path.loc[~data_path.StudyDay.isin([-48,-49,10,12,19,30])]

# Convert days to string (for delta timepoints)
data_path['StudyDay'] = data_path['StudyDay'].astype(str)

# Convert data values to numeric, will error if there are non-numeric data values
data_path.ENTRY = pd.to_numeric(data_path.ENTRY)

### Reshape table

# Not including ANALYSIS and SAMPLE_NUMBER so the rows are merged
# Not including for simplicity: DOB, SAMPLED_DATE, CLASS_CODE, and 'SPECIES_STRAIN'
id_cols_path = ['STUDY_NO','Class','Subject','SEX','COUNTRY_ORIGIN','StudyDay','Age_d']

# Check that the pivot won't be averaging any rows together by outputting the count instead of the mean (All values should be 1)
test = data_path.pivot_table(values='ENTRY',columns='NAME',index=id_cols_path,aggfunc=lambda x: len(x.unique()))
var_cols_path = test.columns.tolist() # save for later
# Check for any values over 1
ninetynine(test[test > 1].any().any(),'merging values across rows')

# Perform table pivot
data_path = data_path.pivot_table(values='ENTRY',columns='NAME',index=id_cols_path)
data_path = data_path.reset_index()
data_path.head()

# check for duplicates after the table reshaping
ninetynine(len(data_path[data_path.duplicated(subset=['Subject','StudyDay'])]) > 0,'duplicate values')

### Calculate delta terms

# Create temporary dataframe to add the delta timepoints
data_path = data_path.set_index(['Subject','StudyDay']).unstack()

for var in var_cols_path:
    # Collapse pre-exposure values (should have one per subject)
    data_path.loc[:,(var,'pre')] = data_path.loc[:,(var,['-11','-7'])].mean(axis = 1, skipna = True)

    # Calculate the change from pre-exposure
    data_path.loc[:,(var,'pre_delta')] = data_path.loc[:,(var,'pre')]-data_path.loc[:,(var,'pre')] #should be all 0
    data_path.loc[:,(var,'2_delta')] = data_path.loc[:,(var,'2')]-data_path.loc[:,(var,'pre')]
    data_path.loc[:,(var,'4_delta')] = data_path.loc[:,(var,'4')]-data_path.loc[:,(var,'pre')]
    data_path.loc[:,(var,'6_delta')] = data_path.loc[:,(var,'6')]-data_path.loc[:,(var,'pre')]
    data_path.loc[:,(var,'8_delta')] = data_path.loc[:,(var,'8')]-data_path.loc[:,(var,'pre')]

# Copy the class and age to the new columns from the original columns

# stack and unstack so that missing columns fill in with null
data_path = data_path.stack().unstack()

# Copy class to other time points
time_points=data_path.loc[:,'Class'].columns.to_list()
# back-fill and foward-fill incase the column order changes
data_path.loc[:,('Class',time_points)] = data_path.loc[:,('Class',time_points)].fillna(method='ffill',axis=1).fillna(method='bfill',axis=1)

# Copy age to delta time point
tmp_days = ['2','2_delta','4','4_delta','6','6_delta','8','8_delta']
data_path.loc[:,('Age_d',tmp_days)] = data_path.loc[:,('Age_d',tmp_days)].fillna(method='ffill',axis=1,limit=1)
data_path.loc[:,('Age_d','pre')] = data_path.loc[:,('Age_d',['-7','-11'])].mean(axis=1)
data_path.loc[:,('Age_d','pre_delta')] = data_path.loc[:,('Age_d',['-7','-11'])].mean(axis=1)

# Reformat for readability
data_path = data_path.stack()

### Save tables

# Save full table (no error values)
data_path.to_csv(os.path.join(PATH,'tables','data_path.csv'))

# Save simplified table (only delta timepoints)
#data_path = data_path.reset_index()
data_path_delta = data_path.reset_index().drop(data_path.reset_index()[data_path.reset_index()['StudyDay'].isin(['-11','-7','2','4','6','8','pre','pre_delta'])].index)
data_path_delta.set_index(['Subject','StudyDay']).to_csv(os.path.join(PATH,'tables','data_path_delta.csv'))

### Exclude variables
with missing values and from domain-specific feature screening

# Uncomment to view variables to be excluded
#data_path[['C-Reactive_Protein','Class']].groupby(['StudyDay','Class']).describe()
#data_path[['Creatinine_Kinase','Class']].groupby(['StudyDay','Class']).describe()
#data_path[['ICT','Class']].groupby(['StudyDay','Class']).describe()

# Variables with missing values
var_cols_path.remove('C-Reactive_Protein')
    # Missing from most subjects
var_cols_path.remove('Creatinine_Kinase')
    # Confirmed missing G57L day 6 from raw
var_cols_path.remove('ICT')

# Variables removed by recommendation of TC
var_cols_exclude = ['Percent_Basophil',
                    'Percent_Eosinophil',
                    'Percent_Lymphocyte',
                    'Percent_Monocyte',
                    'Percent_Neutrophil',
                    'Percent_Reticulocyte']

var_cols_warning = ['Mean_Platelet_Volume_',
                    'Platelet_Distribution_Width',
                    'Platelet_Large_Cell_-_Ratio',
                    'Globulin',
                    'HEM',
                    'LIP']

for var in var_cols_exclude:
    var_cols_path.remove(var)

for var in var_cols_warning:
    var_cols_path.remove(var)

#Write status message, change this anytime this part of the analysis changes
status='CoopExcl'
with open(os.path.join('..','config','analysis_status','q_var_path.txt'),'w') as out_file:
    out_file.write(status)

# save lists
with open('../config/lists_path.pkl', 'wb') as f:
    pickle.dump([id_cols_path,var_cols_path], f)

print(len(var_cols_path))
var_cols_path

### Useful views of the data

# Useful views of the data for troubleshooting error values

#id_cols_path = ['SAMPLE_NUMBER','STUDY_NO','CLASS_CODE','Class','Subject','SEX','DOB','SPECIES_STRAIN','COUNTRY_ORIGIN','StudyDay','SAMPLED_DATE'
#data_path.iloc[5354]
#data_path.NAME.unique()
#data_path.loc[(data_path.NAME == 'C-Reactive Protein') & (data_path.ENTRY == '<0.1'),'ENTRY'] = 0
#data_path.loc[data_path.NAME == 'Creatinine Kinase','ENTRY'].unique()
#data_path.describe()
#data_path.loc[data_path.Subject == 'B03757',['STUDY_NO','CLASS_CODE','Class','Subject','SEX','DOB','SPECIES_STRAIN','COUNTRY_ORIGIN','StudyDay','SAMPLED_DATE','ANALYSIS']]
#data_path.loc[data_path.Subject == 'H54R',id_cols_path]
#data_path.StudyDay.unique()

#len(data_path.loc[(data_path.Subject == 'B03757') & (data_path.ANALYSIS == 'CHEM_CATALYSTONE')].dropna(how='all',axis=1).columns)
#len(data_path.loc[(data_path.Subject == 'B03757') & (data_path.ANALYSIS == 'CBC_PROCYTEDX')].dropna(how='all',axis=1).columns)
#len(data_path.loc[data_path.Subject == 'B03757'].dropna(how='all',axis=1).columns)
#len(data_path.loc[data_path.Subject == 'B03757'].columns)
#all_col = data_path.columns.to_list().remove(id_cols_path)
#all_sanNA = data_path.loc[data_path.Subject == 'B03757'].dropna(how='all',axis=1).columns

#data_path.describe().loc['count',:]
#data_path.info()

#data_path.groupby(['StudyDay','Class']).agg(['count','mean','std','min','max'])

#data_path.describe()

#data_path.reset_index().loc[data_path.reset_index().Subject == 'G57L',['StudyDay','Creatinine_Kinase']]

#data_path.loc[(data_path.Class == 'Mock') & (data_path.StudyDay == '2')]

#data_path.reset_index().loc[data_path.reset_index().Class == 'Mock','Subject'].unique()
# data_path[['Creatinine_']].loc['H89K']

data_path.groupby(['StudyDay','Class','SEX','COUNTRY_ORIGIN']).agg(['count','mean','std','min','max'])

# data_path.reset_index().loc[~data_path.reset_index().Subject.isin(['H56N', 'H89K','B03942','G21D'])].groupby(['StudyDay','Class','SEX','COUNTRY_ORIGIN']).agg(['count','mean','std','min','max'])

#data_path.reset_index().loc[(data_path.reset_index().SEX == 'F') & (data_path.reset_index().Class == 'Mock'),'Subject'].unique()

### Notes
* Columns removed because only contains data >20 days pre-infection (and for only 3 subjects): 'Adjusted Ca', 'Calcium', 'Gamma-glutamyltransferase', 'Potassium', 'Sodium', 'Total Bilirubin', 'Triglyceride'
* C-Reactive Protein was only collected for 64E-1 so that is why there are only 47 entries for this modality
