==> notebooks/01c_Data_Cleaning_immune.ipynb <==
import pandas as pd
import pickle
import yaml
import os
import sys

# Local packages
sys.path.append('../src')
from ninetynine import ninetynine

#open data
with open('../config/paths.yaml','r') as file:
    paths_list = yaml.safe_load(file)
    PATH = os.path.abspath(paths_list['PATH'])

data_imm = pd.read_csv(os.path.join(PATH,'raw','GPP_Final_Output_allTime.csv'))
data_imm.head()

# Views of the raw data
#data_imm.astype('object').describe()
#len(data_imm[data_imm.NAME == 'C-Reactive Protein'])
#data_imm.loc[data_imm.NAME == 'C-Reactive Protein',['SAMPLE_NUMBER','SUBJECT_ID','STUDY_DAY','STUDY_NO']]
#data_imm.loc[data_imm.NAME == 'ICT',['SAMPLE_NUMBER','SUBJECT_ID','STUDY_DAY','STUDY_NO','ENTRY']].astype(object).describe()
#data_imm.info()
#data_imm.loc[data_imm['SUBJECT_ID'] == 'B03757','NAME']
#data_imm.loc[(data_imm['SUBJECT_ID'] == 'G21E') & (data_imm['NAME'] == 'Hemoglobin')]

### Rename columns and values

# Rename columns and values to follow previous convention
data_imm.rename(columns = {'SUBJECT_ID':'Subject','CLASS_DESCRIPTOR':'Class','STUDY_DAY':'StudyDay'},inplace=True)
data_imm.loc[data_imm.Class == 'Virus','Class'] = 'Infected'

# Get rid of spaces and slashes, these are annoying later
data_imm['NAME'] = data_imm['NAME'].replace(' ','_',regex=True)
data_imm['NAME'] = data_imm['NAME'].replace('/','v',regex=True)

### Reshape table

id_cols_imm = ['STUDY_NO','Class','Subject','StudyDay']

# Check that the pivot won't be averaging any rows together by outputting the count instead of the mean (All values should be 1)
test = data_imm.pivot_table(values='ENTRY',columns='NAME',index=id_cols_imm,aggfunc=lambda x: len(x.unique()))
var_cols_imm = test.columns.tolist() # save for later
# Check for any values over 1
ninetynine(test[test > 1].any().any(),'merging values across rows')

# Perform table pivot
data_imm = data_imm.pivot_table(values='ENTRY',columns='NAME',index=id_cols_imm)
data_imm = data_imm.reset_index()
data_imm.head()

# check for duplicates after the table reshaping
ninetynine(len(data_imm[data_imm.duplicated(subset=['Subject','StudyDay'])]) > 0, 'duplicate values')

### Calculate delta terms

# Create temporary dataframe to add the delta timepoints
data_imm = data_imm.set_index(['Subject','StudyDay']).unstack()

for var in var_cols_imm:
    # Collapse the pre-exposure values (should have one per subject)
    data_imm.loc[:,(var,'pre')] = data_imm.loc[:,(var,['-11','-7','B1'])].mean(axis = 1, skipna = True)
    test = data_imm.loc[:,(var,['-11','-7','B1'])].count(axis = 1)
    # Calculate the change from pre-exposure
    data_imm.loc[:,(var,'pre_delta')] = data_imm.loc[:,(var,'pre')]-data_imm.loc[:,(var,'pre')] #should be all 0
    data_imm.loc[:,(var,'2_delta')] = data_imm.loc[:,(var,'2')]-data_imm.loc[:,(var,'pre')]
    data_imm.loc[:,(var,'4_delta')] = data_imm.loc[:,(var,'4')]-data_imm.loc[:,(var,'pre')]
    data_imm.loc[:,(var,'6_delta')] = data_imm.loc[:,(var,'6')]-data_imm.loc[:,(var,'pre')]
    data_imm.loc[:,(var,'8_delta')] = data_imm.loc[:,(var,'8')]-data_imm.loc[:,(var,'pre')]

# Check for any values over 1
ninetynine(test[test > 1].any().any(),'multiple baseline values')

# stack and unstack so that missing columns fill in with null
data_imm = data_imm.stack().unstack()

# Copy class to other time points
time_points=data_imm.loc[:,'Class'].columns.to_list()
# back-fill and foward-fill incase the column order changes
data_imm.loc[:,('Class',time_points)] = data_imm.loc[:,('Class',time_points)].fillna(method='ffill',axis=1).fillna(method='bfill',axis=1)

# Reformat for readability
data_imm = data_imm.stack()

### Useful views of the data

# Useful views of the data for troubleshooting error values

#id_cols_imm = ['SAMPLE_NUMBER','STUDY_NO','CLASS_CODE','Class','Subject','SEX','DOB','SPECIES_STRAIN','COUNTRY_ORIGIN','StudyDay','SAMPLED_DATE'
#data_imm.iloc[5354]
#data_imm.NAME.unique()
#data_imm.loc[(data_imm.NAME == 'C-Reactive Protein') & (data_imm.ENTRY == '<0.1'),'ENTRY'] = 0
#data_imm.loc[data_imm.NAME == 'Creatinine Kinase','ENTRY'].unique()
#data_imm.describe()
#data_imm.loc[data_imm.Subject == 'B03757',['STUDY_NO','CLASS_CODE','Class','Subject','SEX','DOB','SPECIES_STRAIN','COUNTRY_ORIGIN','StudyDay','SAMPLED_DATE','ANALYSIS']]
#data_imm.loc[data_imm.Subject == 'H54R',id_cols_imm]
#data_imm.StudyDay.unique()

#len(data_imm.loc[(data_imm.Subject == 'B03757') & (data_imm.ANALYSIS == 'CHEM_CATALYSTONE')].dropna(how='all',axis=1).columns)
#len(data_imm.loc[(data_imm.Subject == 'B03757') & (data_imm.ANALYSIS == 'CBC_PROCYTEDX')].dropna(how='all',axis=1).columns)
#len(data_imm.loc[data_imm.Subject == 'B03757'].dropna(how='all',axis=1).columns)
#len(data_imm.loc[data_imm.Subject == 'B03757'].columns)
#all_col = data_imm.columns.to_list().remove(id_cols_imm)
#all_sanNA = data_imm.loc[data_imm.Subject == 'B03757'].dropna(how='all',axis=1).columns

#data_imm.describe().loc['count',:]
#data_imm.info()

data_imm.groupby(['StudyDay','Class']).agg(['count','mean','std','min','max'])

#data_imm.describe()
#data_imm.loc[data_imm.Subject == 'H54R']

#data_imm.loc[(data_imm.Class == 'Mock') & (data_imm.StudyDay == '2')]

# Save full table (no error values)
data_imm.to_csv(os.path.join(PATH,'tables','data_imm.csv'))

# Save simplified table (only delta timepoints)
data_imm = data_imm.reset_index()
data_imm_delta = data_imm.drop(data_imm[data_imm['StudyDay'].isin(['10','12','19','30','B1','NEC','-11','-7','2','4','6','8','pre','pre_delta'])].index)
data_imm_delta.set_index(['Subject','StudyDay']).to_csv(os.path.join(PATH,'tables','data_imm_delta.csv'))

print('-- Count of NA entries for each column --')
for var in var_cols_imm:
    na_count = data_imm_delta.loc[data_imm_delta.StudyDay.isin(['2_delta','4_delta','6_delta']),var].isna().sum()
    if na_count > 0:
        ninetynine(na_count > 0,'Missing values ('+var+' - '+str(na_count)+')')

# Remove variables with missing values
var_cols_imm.remove('G-CSF')
var_cols_imm.remove('TGF-alpha')

var_cols_imm

# save lists
with open('../config/lists_imm.pkl', 'wb') as f:
    pickle.dump([id_cols_imm,var_cols_imm], f)

data_imm_delta
