==> notebooks/01d_Data_Cleaning_merge.ipynb <==
import os
import pandas as pd
import yaml
import pickle
import sys

# Local packages
sys.path.append('../src')
from ninetynine import ninetynine

#Retreive local data path
with open('../config/paths.yaml','r') as file:
    paths_list = yaml.safe_load(file)
    PATH = os.path.abspath(paths_list['PATH'])

# Retreive s_list
with open('../config/s_list.yaml','r') as file:
    s_list = yaml.safe_load(file)['s_list']

# Getting back the objects:
with open('../config/lists_radio.pkl', 'rb') as f:
    id_cols_radio,var_cols_radio,time_points,delta_time = pickle.load(f)

# Open cleaned data
data_radio = pd.read_csv(os.path.join(PATH,'tables','data_radio_delta.csv'))
data_radio.head()

# Getting back the objects:
with open('../config/lists_path.pkl', 'rb') as f:
    id_cols_path,var_cols_path = pickle.load(f)

# Open cleaned data
data_path = pd.read_csv(os.path.join(PATH,'tables','data_path_delta.csv'))
data_path.head()

# Getting back the objects:
with open('../config/lists_imm.pkl', 'rb') as f:
    id_cols_imm,var_cols_imm = pickle.load(f)

# Open cleaned data
data_imm = pd.read_csv(os.path.join(PATH,'tables','data_imm_delta.csv'))
data_imm.head()

id_cols = id_cols_radio + id_cols_path + id_cols_imm
var_cols = var_cols_radio + var_cols_path + var_cols_imm

data_merge = data_path.set_index(['Subject','StudyDay']).join(data_radio.set_index(['Subject','StudyDay']),lsuffix='_path', rsuffix='_radio').reset_index()
data_merge

data_merge = data_merge.set_index(['Subject','StudyDay']).join(data_imm.set_index(['Subject','StudyDay']),rsuffix='_imm').reset_index()
data_merge

# Check if the two datasets have matching values
ninetynine(not data_merge['Class_radio'].equals(data_merge['Class_path']),'mis-matched class labels')
ninetynine(not data_merge['Class_radio'].equals(data_merge['Class']),'mis-matched class labels')
#data_merge.loc[data_merge.Class_radio != data_merge.Class_path,['Class_path','Class_radio']]

# Create a single Class field
data_merge['Class_name'] = data_merge['Class_path']

# Enumerate Class
data_merge.loc[data_merge["Class_name"] == "Infected","Class"] = 1
data_merge.loc[data_merge["Class_name"] == "Mock","Class"] = 0

#### Check and save results

data_merge = data_merge.loc[data_merge.Subject.isin(s_list)]
print('Subject count: ',len(data_merge.Subject.unique()))

#Write status message, change this anytime this part of the analysis changes
status='SexOriginBal'
with open(os.path.join('..','config','analysis_status','t_merge.txt'),'w') as out_file:
    out_file.write(status)

for i in data_merge.columns:
    print(i)

data_merge['StudyDay'].unique()

data_merge.set_index(['Subject','StudyDay']).to_csv(os.path.join(PATH,'tables','data_merge.csv'))

data_merge_246 = data_merge.loc[data_merge.StudyDay != '8_delta']
data_merge_246.set_index(['Subject','StudyDay']).to_csv(os.path.join(PATH,'tables','data_merge246.csv'))

#Check for any NA values in the 246 table, cannot have NA values for future analyses
for var in var_cols:
    if data_merge_246[var].isna().any():
        print(data_merge_246.loc[data_merge_246[var].isna(),[var,'Subject','StudyDay']])

id_cols
