==> notebooks/02b_Demographic_info.ipynb <==
import os
import sys
import pandas as pd
import yaml

import scipy.stats

#Retreive local data path
with open('../config/paths.yaml','r') as file:
    paths_list = yaml.safe_load(file)
    PATH = os.path.abspath(paths_list['PATH'])

data = pd.read_csv(os.path.join(PATH,'tables','data_merge_noDelta.csv'))
data.groupby(['StudyDay','Class_name']).describe()

data_clinScore = pd.read_csv(os.path.join(PATH,'raw','Cage-side_scores_v2.csv'))

data_clinScore = data_clinScore.pivot_table(index = 'Subject', values = 'Clinical_score', columns = 'StudyDay')
data_clinScore.columns = data_clinScore.columns.astype(str)

data_clinScore

# Calculate the change from pre-exposure
data_clinScore['2_delta'] = data_clinScore['2'] - data_clinScore['-7']
data_clinScore['4_delta'] = data_clinScore['4'] - data_clinScore['-7']
data_clinScore['6_delta'] = data_clinScore['6'] - data_clinScore['-7']
data_clinScore['max'] = data_clinScore[['2','4','6']].max(axis=1)

# get the baseline age in years
data['Age_bl_y'] = (data.loc[data.StudyDay==2,'Age_d']-2)/365

data_clin_score = data[['Subject','Class_name','SEX','COUNTRY_ORIGIN','Age_bl_y']].dropna().drop_duplicates(subset=None).set_index('Subject').join(data_clinScore[['max']])
data_clin_score

data_clin_score.groupby('Class_name').describe()

def welch_dof(x,y):
        dof = (x.var()/x.size + y.var()/y.size)**2 / ((x.var()/x.size)**2 / (x.size-1) + (y.var()/y.size)**2 / (y.size-1))
        print(f"Welch-Satterthwaite Degrees of Freedom= {dof:.4f}")

scipy.stats.ttest_ind(data_clin_score.loc[data_clin_score.Class_name == 'Infected','max'],data_clin_score.loc[data_clin_score.Class_name == 'Mock','max'], equal_var=False)

welch_dof(data_clin_score.loc[data_clin_score.Class_name == 'Infected','max'], data_clin_score.loc[data_clin_score.Class_name == 'Mock','max'])

scipy.stats.ttest_ind(data_clin_score.loc[data_clin_score.Class_name == 'Infected','Age_bl_y'],data_clin_score.loc[data_clin_score.Class_name == 'Mock','Age_bl_y'], equal_var=False)

welch_dof(data_clin_score.loc[data_clin_score.Class_name == 'Infected','Age_bl_y'], data_clin_score.loc[data_clin_score.Class_name == 'Mock','Age_bl_y'])

data_clin_score.index.tolist()

for col in data.columns:
    print(col)

data_infectMode = pd.read_csv(os.path.join(PATH,'raw','Cage-side_scores_v2.csv'),dtype={'Study':str,'Group':str})

#data_infectMode = data_infectMode.pivot_table(index = 'Subject', values = 'Group ', columns = 'StudyDay', aggfunc=lambda x: ' '.join(x))
#data_infectMode.columns = data_infectMode.columns.astype(str)

data_infectMode.loc[(data_infectMode.Study == '64E-1a'),'mode'] = 'IB'
data_infectMode.loc[(data_infectMode.Study == '64E-2a'),'mode'] = 'IB'
data_infectMode.loc[((data_infectMode.Group == '1') | (data_infectMode.Group == '2')) & (data_infectMode.Study == '64E-3a'),'mode'] = 'IB'
data_infectMode.loc[((data_infectMode.Group == '3') | (data_infectMode.Group == '4')) & (data_infectMode.Study == '64E-3a'),'mode'] = 'Aerosol'

data_infectMode.loc[(data_infectMode.Group == '1') & (data_infectMode.Study == '64E-1a'),'Class_name'] = 'Mock'
data_infectMode.loc[(data_infectMode.Group == '2') & (data_infectMode.Study == '64E-1a'),'Class_name'] = 'Infected'
data_infectMode.loc[(data_infectMode.Group == '1') & (data_infectMode.Study == '64E-2a'),'Class_name'] = 'Mock'
data_infectMode.loc[(data_infectMode.Group == '2') & (data_infectMode.Study == '64E-2a'),'Class_name'] = 'Infected'
data_infectMode.loc[((data_infectMode.Group == '1') | (data_infectMode.Group == '3')) & (data_infectMode.Study == '64E-3a'),'Class_name'] = 'Mock'
data_infectMode.loc[((data_infectMode.Group == '2') | (data_infectMode.Group == '4')) & (data_infectMode.Study == '64E-3a'),'Class_name'] = 'Infected'

data_infectMode.loc[data_infectMode.Subject.isin(data_clin_score.index.tolist())].drop_duplicates('Subject').groupby(['Class_name','mode']).agg('count')

data_infectMode.loc[data_infectMode.Subject.isin(data_clin_score.index.tolist())].drop_duplicates('Subject').groupby(['Class_name','Sex']).agg('count')

data_infectMode.loc[data_infectMode.Subject.isin(data_clin_score.index.tolist())].drop_duplicates('Subject').groupby(['Class_name','mode']).agg('count')

data_infectMode

testchi = data_infectMode.loc[data_infectMode.Subject.isin(data_clin_score.index.tolist())].drop_duplicates('Subject').groupby(['mode','Class_name']).agg('count')[['Subject']].unstack()

display(testchi)
chi2, chiP, dof, expected = scipy.stats.chi2_contingency(testchi)
print('Chi2: ',chi2)
print('dof: ',dof)
print('pval: ',chiP)

testchi = data_infectMode.loc[data_infectMode.Subject.isin(data_clin_score.index.tolist())].drop_duplicates('Subject').groupby(['Sex','Class_name']).agg('count')[['Subject']].unstack()

display(testchi)
chi2, chiP, dof, expected = scipy.stats.chi2_contingency(testchi)
print('Chi2: ',chi2)
print('dof: ',dof)
print('pval: ',chiP)

# Get weight demographic info

data_weight = pd.read_csv(os.path.join(PATH,'raw','64E_weights_compiled_v2.csv'),dtype={'Study':str,'Group':str})

#data_weight = data_weight.pivot_table(index = 'Subject', values = 'Group ', columns = 'StudyDay', aggfunc=lambda x: ' '.join(x))
#data_weight.columns = data_weight.columns.astype(str)

data_weight.loc[(data_weight.Study == '64E-1a'),'mode'] = 'IB'
data_weight.loc[(data_weight.Study == '64E-2a'),'mode'] = 'IB'
data_weight.loc[((data_weight.Group == '1') | (data_weight.Group == '2')) & (data_weight.Study == '64E-3a'),'mode'] = 'IB'
data_weight.loc[((data_weight.Group == '3') | (data_weight.Group == '4')) & (data_weight.Study == '64E-3a'),'mode'] = 'Aerosol'

data_weight.loc[(data_weight.Group == '1') & (data_weight.Study == '64E-1a'),'Class_name'] = 'Mock'
data_weight.loc[(data_weight.Group == '2') & (data_weight.Study == '64E-1a'),'Class_name'] = 'Infected'
data_weight.loc[(data_weight.Group == '1') & (data_weight.Study == '64E-2a'),'Class_name'] = 'Mock'
data_weight.loc[(data_weight.Group == '2') & (data_weight.Study == '64E-2a'),'Class_name'] = 'Infected'
data_weight.loc[((data_weight.Group == '1') | (data_weight.Group == '3')) & (data_weight.Study == '64E-3a'),'Class_name'] = 'Mock'
data_weight.loc[((data_weight.Group == '2') | (data_weight.Group == '4')) & (data_weight.Study == '64E-3a'),'Class_name'] = 'Infected'

data_weight = data_weight.set_index('Animal ID')

data_weight['BL'] = data_weight[['D-11','D-7']].mean(axis=1)
data_weight['delta_2'] = data_weight.D2 - data_weight.BL
data_weight['delta_4'] = data_weight.D4 - data_weight.BL
data_weight['delta_6'] = data_weight.D6 - data_weight.BL

data_weight['delta_mean'] = data_weight[['delta_2','delta_4','delta_6']].mean(axis=1)

data_weight = data_weight.loc[data_weight.index.isin(data_clin_score.index.tolist())]

data_weight.groupby('Class_name').describe().delta_mean

scipy.stats.ttest_ind(data_weight.loc[data_weight.Class_name == 'Infected','delta_mean'], data_weight.loc[data_weight.Class_name == 'Mock','delta_mean'],equal_var=False)

welch_dof(data_weight.loc[data_weight.Class_name == 'Infected','delta_mean'], data_weight.loc[data_weight.Class_name == 'Mock','delta_mean'])

data_weight.groupby('Class_name').describe().BL

scipy.stats.ttest_ind(data_weight.loc[data_weight.Class_name == 'Infected','BL'], data_weight.loc[data_weight.Class_name == 'Mock','BL'],equal_var=False)

welch_dof(data_weight.loc[data_weight.Class_name == 'Infected','BL'], data_weight.loc[data_weight.Class_name == 'Mock','BL'])
