==> notebooks/03_Feature_Selection.ipynb <==
import os
import yaml
import pandas as pd
import pickle
import matplotlib.pyplot as plt

import pymrmr
import time

import sys
sys.path.append('../src')

from sklearn.preprocessing import StandardScaler
from sklearn.feature_selection import GenericUnivariateSelect, chi2, f_classif, mutual_info_classif
from sklearn.preprocessing import MinMaxScaler

# Getting back the objects:
with open('../config/lists_radio.pkl', 'rb') as f:
    id_cols_radio,var_cols_radio,time_points,delta_time = pickle.load(f)

with open('../config/lists_path.pkl', 'rb') as f:
    id_cols_path,var_cols_path = pickle.load(f)

with open('../config/lists_imm.pkl', 'rb') as f:
    id_cols_imm,var_cols_imm = pickle.load(f)

with open('../config/lists_classEffect.pkl', 'rb') as f:
    var_cols_classEffect = pickle.load(f)
    
id_cols = id_cols_radio + id_cols_path + id_cols_imm + ['Class_name']
var_cols = var_cols_radio + var_cols_path + var_cols_imm
#var_cols = var_cols_imm + var_cols_path + var_cols_radio
    
#Retreive local data path
with open('../config/paths.yaml','r') as file:
    paths_list = yaml.safe_load(file)
    PATH = os.path.abspath(paths_list['PATH'])
    
# Open cleaned data
data = pd.read_csv(os.path.join(PATH,'tables','data_merge246.csv'))
data.head()

X = data[var_cols]
y = data.Class

def show_list(l):
    i=0
    for t in l:
        i += 1
        print(str(i)+') '+t)

def get_modality(name):
    if name in var_cols_radio:
        return 'Radiomics'
    elif name in var_cols_path:
        return 'Clinical Pathology'
    elif name in var_cols_imm:
        return 'Immunology'

show_list(var_cols)

### Feature Selection: test different scoring functions

(Lists are not ranked)

k=34

mask = GenericUnivariateSelect(score_func=f_classif,mode='k_best', param=k).fit(X, y).get_support()
var_cols_fbestk = X.columns[mask].to_list()
print('----- f-stat -----')
#show_list(var_cols_fbestk)

mask = GenericUnivariateSelect(score_func=mutual_info_classif,mode='k_best', param=k).fit(X, y).get_support()
var_cols_MIbestk = X.columns[mask].to_list()
print('----- Mutual Information -----')
show_list(var_cols_MIbestk)

X_scaled = MinMaxScaler().fit(X).transform(X) #scale to 0-1
mask = GenericUnivariateSelect(score_func=chi2,mode='k_best', param=k).fit(X_scaled, y).get_support()
var_cols_chibestk = X.columns[mask].to_list()
print('----- Chi-squared -----')
#show_list(var_cols_chibestk)

var_cols_intersectbestk = list(set(var_cols_fbestk) & set(var_cols_MIbestk) & set(var_cols_chibestk))
print('----- f, MI, chi intersection -----')
#show_list(var_cols_intersectbestk)

### Feature Selection: Percentile (10%)

# mask = GenericUnivariateSelect(score_func=f_classif,mode='fpr', param=0.05).fit(X, y).get_support()
# var_cols_MIbestk = X.columns[mask].to_list()
# print('----- False-positive rate -----')
# show_list(var_cols_MIbestk)

# mask = GenericUnivariateSelect(score_func=f_classif,mode='percentile', param=10).fit(X, y).get_support()
# var_cols_MIbestk = X.columns[mask].to_list()
# print('----- Mutual Information -----')
# show_list(var_cols_MIbestk)

### MRMR

#z-score normalize
data_znorm = data.copy()

for var in var_cols:
    data_znorm[var] = (data_znorm[var]-data_znorm[var].mean()) / data_znorm[var].std()

data_znorm.describe()

# var_cols = var_cols_radio + var_cols_path

# var_cols.remove('Blood_Urea_Nitrogen')
# var_cols.remove('Creatinine_')

#data_znorm.isna().any()

for var in var_cols:
    if data_znorm[var].isna().any():
        print(var)

### MRMR parameter definition

if False:
    # Choose either Mutual information quotient (MIQ) or Mutual information difference (MID)
    # Github is here: https://github.com/fbrundu/pymrmr

    k_tmp=35#len(data)/2
    #k_tmp=35
    print('k = ',k_tmp)
    #var_cols_mrmr = pymrmr.mRMR(data_znorm[['Class']+var_cols],'MIQ', k_tmp)
    var_cols_mrmr = pymrmr.mRMR(data_znorm[['Class']+var_cols],'MID', k_tmp)
    var_cols_mrmr
    print('Radiomics',[get_modality(n) for n in var_cols_mrmr].count('Radiomics'))
    print('Clinical Pathology',[get_modality(n) for n in var_cols_mrmr].count('Clinical Pathology'))
    print('Immunology',[get_modality(n) for n in var_cols_mrmr].count('Immunology'))
    
if False:
    tmp_df = pd.DataFrame({'var':var_cols_mrmr,'mod':[get_modality(n) for n in var_cols_mrmr]})
    tmp_df.to_csv(os.path.join(PATH,'tables','feature_selection_mrmr.csv'))
    print(os.path.join(PATH,'tables','feature_list_mrmr.csv'))

### MRMR Permute Call

recompute_mrmr_permute = False

if recompute_mrmr_permute:
    from ml_classic_repo import mrmr_permute
    corr_mrmr = mrmr_permute.mrmr_permute(data_znorm = data_znorm,var_cols = var_cols,a_thresh = 0.05)
    corr_mrmr.to_csv(os.path.join(PATH,'tables','mrmr_permute_05_20211020_demo.csv'))
else:
    corr_mrmr = pd.read_csv(os.path.join(PATH,'tables','mrmr_permute_05_20211013.csv'))

corr_mrmr['Modality'] = [get_modality(n) for n in corr_mrmr['Name'].tolist()]
    
display(corr_mrmr) #.Name.values.tolist()
#print('k='+str(len(corr_mrmr)))

#k_mp = len(corr_mrmr)

#stop_right_here

#corr_mrmr.to_csv(os.path.join(PATH,'tables','mrmr_permute005b.csv'))

### This is where var_cols_FS is defined

# --- Define var_cols_FS ---
#var_cols_FS = var_cols_fbestk
#var_cols_FS = var_cols_intersectbestk
#var_cols_FS=['Idmn','GrayLevelNonUniformity','Uniformity','Entropy','GrayLevelNonUniformityNormalized','LongRunLowGrayLevelEmphasis','ZoneVariance','LargeAreaEmphasis','Busyness']
#var_cols_FS = var_cols_classEffect
#var_cols_FS = var_cols_mrmr
var_cols_FS = corr_mrmr.Name.values.tolist()

# --- Define status ---
#Write status message, change this anytime this part of the analysis changes
#status='FS:2x4ANOVA,Bonferroni'
#status='FS:2x3ANOVA,notCorrected'
#status='FS:2x3ANOVA,fdr'
#status='FS:ftest k=34'
#status='FS:MRMR MID k='+str(k_tmp)
#status='FS:MRMR MIQ k='+str(k_tmp)
status='MRMR-permute 0.05'

with open(os.path.join('..','config','analysis_status','x_feature_selection.txt'),'w') as out_file:
    out_file.write(status)

# ---- save lists ---
with open('../config/lists_FeatureSelect.pkl', 'wb') as f:
    pickle.dump(var_cols_FS, f)

#### Blood_Urea_Nitrogen and Creatinine_ not included in future studies

var_cols_FS

len(var_cols_FS)

anova_table = pd.read_csv(os.path.join(PATH,'tables','anova_join.csv'))
anova_table = corr_mrmr.set_index('Name').join(anova_table.set_index('Unnamed: 0'))
anova_table['Class_name_pval']
#anova_table.to_csv(os.path.join(PATH,'tables','anova_join_mrmr.csv'))

anova_table

### Exploring MRMR

#### MRMR on all measures, top 20, MIQ

pymrmr.mRMR(data_znorm[['Class']+var_cols],'MIQ', 20)

#### MRMR only on radiomics measures

pymrmr.mRMR(data_znorm[['Class']+var_cols_radio],'MIQ', 20)

pymrmr.mRMR(data_znorm[['Class']+var_cols_radio],'MIQ', 20)

#### MRMR only on clinical pathology measures

pymrmr.mRMR(data_znorm[['Class']+var_cols_path],'MIQ', 20)

#### MRMR only on immunology measures

pymrmr.mRMR(data_znorm[['Class']+var_cols_imm],'MIQ', 20)

#### MRMR on all measures MID

pymrmr.mRMR(data_znorm[['Class']+var_cols],'MID', 20)

#### Display MRMR results next to variable type

feature_type=pd.DataFrame()

feature_type['Feature']=pymrmr.mRMR(data_znorm[['Class']+var_cols],'MID', 10)
feature_type['Rank'] = [i + 1 for i in feature_type.index.tolist()]

feature_type.loc[feature_type['Feature'].isin(var_cols_radio),'Type']='Radiomics'
feature_type.loc[feature_type['Feature'].isin(var_cols_path),'Type']='Clinical Pathology'
feature_type.loc[feature_type['Feature'].isin(var_cols_imm),'Type']='Immunology'

feature_type[['Rank','Type','Feature']]

# with open('var_cols_path_list.txt', 'w') as f:
#     for item in var_cols_path:
#         f.write("%s\n" % item)

data_znorm[['Class']+var_cols].set_index('Class').to_csv(os.path.join(PATH,'tables','data_znorm.csv'))

pymrmr.mRMR(data_znorm[['Class']+var_cols],'MID', 10)

# import mrmr_wrapper

# max_rel_df,mRMR_df = mrmr_wrapper.mrmr(data_znorm[['Class']+var_cols].set_index('Class'),'MID', 25)

# mRMR_df = mRMR_df.set_index('Name')
# max_rel_df = max_rel_df.set_index('Name')
# mRMR_df['max_rel'] = max_rel_df.Score
# mRMR_df['redundency'] = mRMR_df.max_rel - mRMR_df.Score
# mRMR_df

# data_znorm[['Class']+var_cols]
