==> notebooks/04b_MRMR_Permute_Example.ipynb <==
import sys
sys.path.append(r'../src')

import pymrmr

from ml_classic_repo import ml_classic

# Analysis functions, dataset specific
def clean(data_tmp):
    # Dataset specific

    #print(ml1.data.groupby(ml1.label).describe())

    #ml1.znorm_all()

    #data_tmp = data_tmp.loc[data_tmp.TARGET.isna()
    data_tmp = data_tmp.dropna()
    return data_tmp


class inputs():
    def __init__(self):
        self.data = '/hpcdata/irf-image/chuwt/classification_data/peng_data/test_colon_s3.csv'
        self.output = '/hpcdata/irf-image/chuwt/classification_data/peng_data/test_colon_s3'
        self.label = 'class'

args = inputs()
    
# # Get data
# data_df = ml_classic.ml_classic(args.data)

# # Clean data
# data_df = clean(data_df)

ml1 = ml_classic.ml_classic(args.data)

ml1.set_label(args.label)

ml1.summary()
#print(ml1.data.groupby(ml1.label).describe())
#print(ml1.data.info())
#ml1.print_columns()
#print(ml1.data.head())

# Run feature selection
max_rel,mrmr_result = ml1.fs_mrmr(n=5)
print(mrmr_result)
#   mrmr_result.to_csv(args.output + '_mrmr.csv')

#   mrmr_permute_result = ml1.fs_mrmr_permute()
#   mrmr_permute_result.to_csv(args.output + '_mrmr_permute.csv')
#   print(mrmr_permute_result)


mrmr_permute_result = ml1.fs_mrmr_permute()

pymrmr.mRMR(ml1.data[[ml1.label]+ml1.features],'MID',5)
