==> notebooks/05_Compare_ML.ipynb <==
import os
import yaml
import pandas as pd
import pickle
import sys
from datetime import datetime

sys.path.append('../src')
from ml_tools_wc import plot_comparison

#Retreive local data path
with open('../config/paths.yaml','r') as file:
    paths_list = yaml.safe_load(file)
    PATH = os.path.abspath(paths_list['PATH'])
    
# Open cleaned data
master_log_file = os.path.join(PATH,'tables','master_log_2021_08_10.csv')
master_log = pd.read_csv(master_log_file)
master_log.head()

master_log.Note.unique()

print('There are duplicates: ',master_log.duplicated(subset=['Note','Model']).any())

# def plot_comparison(log, title_prefix, labelA, noteA, labelB, noteB, metric='CVaccuracy'):
#     model_list=['LogisticRegression()',"SVC(kernel='linear', probability=True)",
#                 'DecisionTreeClassifier()','RandomForestClassifier()',
#                 'KNeighborsClassifier()','GaussianNB()',
#                 'LinearDiscriminantAnalysis()']
#     model_name_dict={"SVC(kernel='linear', probability=True)":'SVM','LogisticRegression()':'Logistic Regression',
#                         'DecisionTreeClassifier()':'Decision Tree','RandomForestClassifier()':'Random Forest',
#                         'KNeighborsClassifier()':'KNN','GaussianNB()':'Gaussian Naive Bayes',
#                         'LinearDiscriminantAnalysis()':'LDA'}
    
#     title = title_prefix + ' ' + labelA + ' vs. ' + labelB
#     tag = 'ML_comparison'+title.replace(' ','_')
#     tmp_log = log.loc[master_log['Note'].isin([noteA,noteB]) & master_log['Model'].isin(model_list),['Model','Note',metric]]
#     tmp_log = tmp_log.replace(model_name_dict)
#     tmp_log.loc[tmp_log.Note == noteA,'Note'] = labelA
#     tmp_log.loc[tmp_log.Note == noteB,'Note'] = labelB
#     tmp_log = tmp_log.set_index(['Model','Note']).unstack()[metric]
#     tmp_log.columns.name = None
#     display(tmp_log)
    
#     fig = plt.figure(facecolor='white')
#     axs = fig.add_subplot()

#     tmp_log.plot.barh(y=[labelA,labelB], ax=axs)
    
#     axs.set_axisbelow(True)
#     axs.xaxis.grid()
#     axs.set_title(title)

#     axs.set_xlim([0.5,1])
#     axs.set_xlabel(metric)
#     fig.savefig(os.path.join(PATH,'figures',tag+'.png'), bbox_inches="tight",dpi=300)
#     plt.show()
#     plt.close(fig)

# plot_comparison(master_log,'MRMR',\
#                 'top 10%',\
#                 'Timepoints d2/d4/d6; FS: MRMR k=17; z-norm; immunology data',\
#                 'top 20%',\
#                 'Timepoints d2/d4/d6; FS: MRMR k=34; z-norm; immunology data')

# plot_comparison(master_log,'',\
#                 'MRMR',\
#                 'Timepoints d2/d4/d6; FS: MRMR k=34; z-norm; immunology data',\
#                 'f-test',\
#                 'Timepoints d2/d4/d6; FS: f-test k=34; z-norm; immunology data')

# plot_comparison(master_log,'',\
#                 'Without BUN or Creatinine',\
#                 'Timepoints d2/d4/d6; FS: MRMR k=34; z-norm; immunology data; no BUN or Creat',\
#                 'With BUN and Creatinine',\
#                 'Timepoints d2/d4/d6; FS: MRMR k=34; z-norm; immunology data')

fig,tag = plot_comparison(master_log,'',\
                'old',\
                'Timepoints d2/d4/d6; FS: MRMR k=34; z-norm; immunology data',\
                'new',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;')
fig.savefig(os.path.join(PATH,'figures',tag+'.png'), bbox_inches="tight",dpi=300)

plot_comparison(master_log,'',\
                'new',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;',\
                'old',\
                'Timepoints d2/d4/d6; FS: MRMR k=34; z-norm; immunology data')

plot_comparison(master_log,'FS comparison ',\
                '2x4ANOVAp>0.05',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:2x4ANOVA;',\
                'MRMR',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;')

plot_comparison(master_log,'No Immunology',\
                'Without BUN and Creatinine',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34,no imm,noBUNorCre;',\
                'With BUN and Creatinine',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34, no imm;')

plot_comparison(master_log,'',\
                'Without BUN or Creatinine',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34,noBUNnoCre;',\
                'With BUN and Creatinine',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;')

plot_comparison(master_log,'MRMR',\
                'top 20%',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;',\
                'top 10%',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=17;')

plot_comparison(master_log,'',\
                'MRMR',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;',\
                'f-test',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:ftest k=34;')

plot_comparison(master_log,'',\
                'MRMR',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;',\
                'All',\
                'time:d2d4d6;z-norm,all_vars;radio-path-imm;FS:MRMR MIQ k=34;')

plot_comparison(master_log,'',\
                'MRMR',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;',\
                'Cooper excl and warn',\
                'TC_ExclAndWarn;time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;')

plot_comparison(master_log,'',\
                'MRMR',\
                'time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;',\
                'Cooper excl',\
                'TC_Excl;time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;')

plot_comparison(master_log,'',\
                'MRMR-MIQ',\
                'TC_ExclAndWarn;time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MIQ k=34;',\
                'MRMR-MID',\
                'TC_ExclAndWarn;time:d2d4d6;z-norm;radio-path-imm;FS:MRMR MID k=34;')
